requirements.txt
fastapi
uvicorn
httpx
pydantic
python-dotenv

langchain
langchain-core
langchain-openai
langgraph
typing-extensions

faiss-cpu
langchain-community

beautifulsoup4
requests
selenium
webdriver-manager

sqlite-utils
SQLAlchemy

aiofiles
jinja2

tqdm
pytest
requests

faiss-cpu